requirements.txt
certifi==2024.7.4
chardet==3.0.4
idna==3.7
matplotlib>=3.4.2
networkx==2.3
numpy>=1.21
opencv-python==4.8.1.78
pillow>=8.3.1
plotly==3.10.0
requests==2.32.2
ipdb==0.13.9
termcolor==1.1.0
tqdm==4.66.3
urllib3>=1.24.3

pyyaml
numpy<2